requirements.txt
# Core application framework
streamlit==1.39.0

# For document processing and Q&A pipeline
# Using modern, compatible versions
langchain==0.2.10
langchain-community==0.2.12
langchain-core==0.2.29
PyMuPDF==1.24.8
sentence-transformers==3.0.1

# For the LLM connection
langchain-groq==0.2.0

# For the vector database
pinecone-client==4.1.1
langchain-pinecone==0.2.0

# Other dependencies
python-dotenv==1.0.1
numpy==1.26.4